moving array.js to utils folder, added David's code changes
