+++ lib/plugins/index.js
@@ -1,8 +1,13 @@
 const { pluginPingPong } = require('./ping-pong')
 const { pluginWelcomeMessage } = require('./welcome-message')
+const { pluginRandomMessage } = require('./random-messages')
 const { PLUGIN_NAME } = require('./symbols')
 
-const plugins = [pluginPingPong, pluginWelcomeMessage]
+const plugins = [
+  pluginPingPong,
+  pluginWelcomeMessage,
+  pluginRandomMessage
+]
 
 async function startPlugins (client) {
   for (const plugin of plugins) {
