+++ lib/utils/array.js
